Started work on removing /tmp properly.
The tmp directory (for r shells) is created by python and R creates
files in it. The rm is ran by nobody, but this means it can't actually
remove the tmp directory itself. This doesn't shut up rm, but Python
will now clean up after it if needed. Hopefully. Not tested yet.